feat: add Dockerfile and use uv as a project manager
