+++ .dockerignore
@@ -1,0 +1,4 @@
+.git
+__pycache__/
+*.pyc
+screens/
