+++ initdb/01_extensions.sql
@@ -1,0 +1,2 @@
+CREATE EXTENSION IF NOT EXISTS postgis;
+CREATE EXTENSION IF NOT EXISTS gevel;
