+++ pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "gist-tree"
+version = "0.1.0"
+requires-python = ">=3.13"
+dependencies = [
+    "pandas>=3.0.1",
+    "psycopg>=3.3.3",
+]
+
+[project.scripts]
+r-tree = "gist_tree.r_tree:main"
+gist-viz = "gist_tree.gist_viz:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gist_tree"]
