+++ src/gist_tree/gist_viz.py
@@ -1,0 +1,161 @@
+"""Display r-tree bboxes"""
+
+import os
+
+from psycopg import connect, sql
+
+from gist_tree.func import unpack
+
+CHECKPOINT = 500  # ask to continue every N inserts
+
+
+def main():
+    conn = connect(
+        host=os.environ.get("PGHOST", "localhost"),
+        dbname=os.environ.get("PGDATABASE", "postgres"),
+        user=os.environ.get("PGUSER", "postgres"),
+        password=os.environ.get("PGPASSWORD", "postgres"),
+        port=os.environ.get("PGPORT", "5432"),
+    )
+    cur = conn.cursor()
+
+    cur.execute(
+        """
+        CREATE TABLE IF NOT EXISTS streets_knn (
+            geom geometry(MultiLineString, 3857));
+            """
+    )
+
+    cur.execute(
+        """
+        TRUNCATE TABLE streets_knn;
+        """
+    )
+
+    for table in ("r_tree_l1", "r_tree_l2"):
+        cur.execute(
+            sql.SQL("CREATE TABLE IF NOT EXISTS {} (geom geometry);").format(
+                sql.Identifier(table)
+            )
+        )
+        cur.execute(
+            sql.SQL("TRUNCATE TABLE {};").format(sql.Identifier(table))
+        )
+
+    cur.execute(
+        """
+        CREATE INDEX IF NOT EXISTS streets_knn_geom_idx
+            ON streets_knn USING gist (geom);
+            """
+    )
+
+    # indices() function must be created beforehand
+    cur.execute(
+        """
+        SELECT index FROM indices() WHERE index = 'streets_knn_geom_idx';
+        """
+    )
+
+    row = cur.fetchone()
+    knn_idx_name = row[0]
+
+    cur.execute(
+        """
+        SELECT COUNT(*) FROM nyc_streets;
+        """
+    )
+    num_geometries = cur.fetchone()[0]
+
+    def bbox(table, scope):
+        cur.execute(
+            sql.SQL(
+                """
+            CREATE TABLE IF NOT EXISTS {} (
+                geom geometry);
+            """
+            ).format(sql.Identifier(table))
+        )
+
+        cur.execute(
+            sql.SQL(
+                """
+            TRUNCATE TABLE {}"""
+            ).format(sql.Identifier(table))
+        )
+
+        cur.execute(
+            sql.SQL(
+                """
+            INSERT INTO {}
+            SELECT st_setsrid(replace(a::text, '2DF', '')::box2d::geometry, 3857)
+            FROM (SELECT * FROM gist_print(%s) as t(level int, valid bool, a box2df) WHERE level = %s) AS subq
+            """
+            ).format(sql.Identifier(table)),
+            [knn_idx_name, scope],
+        )
+
+        cur.execute("NOTIFY qgis;")
+
+    stat = dict()
+
+    for i in range(1, num_geometries + 1):
+        cur.execute(
+            """
+            INSERT INTO streets_knn
+            SELECT
+                n.geom
+            FROM nyc_streets n
+            ORDER by n.geom <-> (
+                SELECT geom FROM nyc_streets
+                WHERE id = 12623)
+            LIMIT 1
+            OFFSET %s;
+            """,
+            [i - 1],
+        )
+
+        conn.commit()
+        conn.autocommit = True
+        cur.execute("VACUUM ANALYZE streets_knn;")
+        conn.autocommit = False
+
+        cur.execute("SELECT gist_stat(%s);", [knn_idx_name])
+
+        for key, value in unpack(cur.fetchone()).items():
+            if key in stat:
+                stat[key].append(value)
+            else:
+                stat[key] = [value]
+
+        for key, value in stat.items():
+            print(f"{key:<16} : {value[i - 1]:,}")
+        print()
+
+        level = stat["Levels"][i - 1]
+
+        level = [value for value in range(1, level + 1)]
+
+        if len(level) == 1:
+            bbox("r_tree_l2", 1)
+
+        else:
+            for j in level:
+                if j == 1:
+                    bbox("r_tree_l1", 1)
+                else:
+                    bbox("r_tree_l2", 2)
+
+        conn.commit()
+
+        if i % CHECKPOINT == 0:
+            answer = (
+                input(f"\n[{i}/{num_geometries}] Continue? (y/n) → ").strip().lower()
+            )
+            if answer != "y":
+                print("Stopping.")
+                break
+
+    conn.commit()
+
+    cur.close()
+    conn.close()
